requirements.txt
Flask==2.3.3
psutil==5.9.0
docker==6.1.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0